src/requirements.txt
about-time==3.1.1
alive-progress==2.1.0
dacite==1.6.0
grapheme==0.6.0
PyYAML==6.0
termcolor==1.1.0
